Alterações nos modelos e provedor do MySQL alterado.

+++ Models/Avaliacao.cs
@@ -7,7 +7,7 @@ namespace Maia.Models
     {
         public Avaliacao(string texto, string titulo, List<string> tags, int usuarioId) : base(texto, titulo, tags, usuarioId) {}
         [Range(0, 10)]
-        public int Nota { get; set; }
+        public byte Nota { get; set; }
         public long? MusicaId { get; set; }
         public Musica Musica { get; set; }
         public int? AlbumId { get; set; }
